requirements.txt
lmdb~=1.2.1
opencv-python~=4.2.0.32
numpy~=1.19.3
torch~=1.7.1
editdistance~=0.5.3
scipy~=1.5.2
torchvision~=0.8.2
imgaug~=0.4.0
Pillow~=7.2.0
six~=1.15.0